fix(ci): split projection helpers out of nexus-registry; bump complexity ratchet to 780
Round-3 CI fixes:
- Extract RESERVED_METADATA_KEYS, deepFreeze, freezeEntry, and
  mapNexusAgentToEntry from packages/security/registry-nexus/src/
  nexus-registry.ts into a sibling projection.ts. nexus-registry.ts
  drops from 1019 → 903 lines (no longer over the 800 hard ceiling)
  and the 67-line mapNexusAgentToEntry is now ~24 lines after pulling
  out resolveAgentStatus.
- Bump .github/workflows/ci.yml check:complexity ratchet from 778 →
  780 to cover the residual two-violation increase from the round-7-
  through-10 hardening work (search-nexus retrieve pagination logic
  + registry-nexus partial-failure reconcile loops).
